Adding Docker to Nuke
I've implemented docker build and docker run with nuke. This will incorporate gitversion to take the image name with whatever the version is. Nuke will generate the docker image with a name of the project and version. These images are saved locally. You can view them in docker dashboard. docker run will run the lastest version by gitversion.

I don't know what port you want to use with docker so you can change the 9000:80 to something else later on. This is tested to work with localhost:9000/swagger and it will take you to the hosted api.

This does require the local machine to have GitVersion 1.0.0.1 and GitVersion.Portable 5.5.0 installed. You can use choco to install these.