Adding Docker to Nuke (#4)
Implemented `BuildDocker` and docker run with Nuke. Incorporated GitVersion to take the image name with whatever the version is and Nuke will generate the docker image with the name of the project and version. Images are saved locally and `DockerRun` will run the latest version by GitVersion. 

The local machine requires to have GitVersion 1.0.0.1 and GitVersion.Portable 5.5.0 installed.